Separate code to .cpp files from .hpp to avoid linker problems
