+++ components/kernel/CMakeLists.txt
@@ -1,7 +1,7 @@
+file(GLOB_RECURSE KERNEL_SRCS "${CMAKE_CURRENT_LIST_DIR}/src/*.cpp")
+
 idf_component_register(
-    SRCS
-        src/drivers/MotorDriver.cpp
-        src/State.cpp
+    SRCS ${KERNEL_SRCS}
     INCLUDE_DIRS
         src
     REQUIRES
@@ -41,7 +41,8 @@ set_target_properties(${COMPONENT_LIB} PROPERTIES WHOLE_ARCHIVE 1)
 
 set(ulp_app_name ulp_pulse_counter)
 set(ulp_sources "${CMAKE_CURRENT_LIST_DIR}/ulp/pulse_counter.c")
-set(ulp_exp_dep_srcs "src/drivers/MotorDriver.cpp" "src/State.cpp")
+# PulseCounter.cpp includes the generated ulp_pulse_counter.h, so it must not
+# be compiled until the ULP binary (and its generated header) are ready.
+set(ulp_exp_dep_srcs "src/PulseCounter.cpp")
 
-include(${IDF_PATH}/components/ulp/cmake/IDFULPComponent.cmake)
 ulp_embed_binary(${ulp_app_name} "${ulp_sources}" "${ulp_exp_dep_srcs}")
