+++ components/kernel/src/PulseCounter.cpp
@@ -1,0 +1,197 @@
+#include "PulseCounter.hpp"
+
+#include <driver/gpio.h>
+#include <esp_pm.h>
+
+#ifdef CONFIG_ULP_COPROC_ENABLED
+#include <driver/rtc_io.h>
+#if defined(CONFIG_ULP_COPROC_TYPE_RISCV)
+#include <ulp_riscv.h>
+#elif defined(CONFIG_ULP_COPROC_TYPE_LP_CORE)
+#include <ulp_lp_core.h>
+#endif
+// Generated by the ULP build system from ulp/pulse_counter.c.
+// The build system prepends "ulp_" to each exported variable name, so
+// channel_count → ulp_channel_count, gpio_num → ulp_gpio_num, etc.
+#include "ulp_pulse_counter.h"
+// Embedded ULP binary produced by CMake.
+extern const uint8_t ulp_pulse_counter_bin_start[] asm("_binary_ulp_pulse_counter_bin_start");
+extern const uint8_t ulp_pulse_counter_bin_end[]   asm("_binary_ulp_pulse_counter_bin_end");
+#endif
+
+#include <EspException.hpp>
+#include <Log.hpp>
+
+using namespace std::chrono;
+
+namespace farmhub::kernel {
+
+// ============================================================================
+// UlpPulseCounter — RTC/LP GPIO path, no per-pulse CPU wakeups
+// ============================================================================
+
+#ifdef CONFIG_ULP_COPROC_ENABLED
+
+/**
+ * @brief Pulse counter backed by the ULP/LP-core coprocessor.
+ *
+ * The coprocessor polls the GPIO continuously regardless of main CPU sleep state:
+ * - ESP32-S3 (ULP-RISC-V): ~17.5 MHz polling rate; requires an RTC-capable GPIO (GPIO 0–21).
+ * - ESP32-C6 (LP Core): ~16 MHz polling rate; requires an LP-capable GPIO (GPIO 0–7).
+ *
+ * The main CPU reads the monotonically increasing counter from RTC slow memory and
+ * computes the delta since the last read. No wakeup sources are registered per pulse.
+ *
+ * PulseCounterManager::start() must be called after all channels are created.
+ *
+ * See docs/ULP-PulseCounter.md for full design rationale.
+ */
+class UlpPulseCounter final : public PulseCounter {
+public:
+    uint32_t getCount() const override {
+        uint32_t current = ulp_pulse_count[channelIndex];
+        uint32_t delta = current - lastSeen;
+        LOGTV(PULSE, "Counted %" PRIu32 " ULP pulses on pin %s",
+            delta, pin->getName().c_str());
+        return delta;
+    }
+
+    uint32_t reset() override {
+        uint32_t current = ulp_pulse_count[channelIndex];
+        uint32_t delta = current - lastSeen;
+        lastSeen = current;
+        LOGTV(PULSE, "Counted %" PRIu32 " ULP pulses and cleared on pin %s",
+            delta, pin->getName().c_str());
+        return delta;
+    }
+
+    PinPtr getPin() const override {
+        return pin;
+    }
+
+private:
+    UlpPulseCounter(InternalPinPtr pin, uint32_t channelIndex)
+        : pin(std::move(pin))
+        , channelIndex(channelIndex)
+        , lastSeen(0) {
+    }
+
+    const InternalPinPtr pin;
+    const uint32_t channelIndex;
+    uint32_t lastSeen;
+
+    friend class PulseCounterManager;
+};
+
+#endif    // CONFIG_ULP_COPROC_ENABLED
+
+// ============================================================================
+// PulseCounterManager
+// ============================================================================
+
+std::shared_ptr<PulseCounter> PulseCounterManager::create(const PulseCounterConfig& config) {
+    auto gpio = config.pin->getGpio();
+
+#ifdef CONFIG_ULP_COPROC_ENABLED
+    if (rtc_gpio_is_valid_gpio(gpio)) {
+        return createUlp(config);
+    }
+#endif
+    return createGpio(config);
+}
+
+void PulseCounterManager::start() {
+#ifdef CONFIG_ULP_COPROC_ENABLED
+    if (ulpStarted || ulpNextChannel == 0) {
+        return;
+    }
+    ulpStarted = true;
+
+    ulp_channel_count = ulpNextChannel;
+
+    size_t size = ulp_pulse_counter_bin_end - ulp_pulse_counter_bin_start;
+#if defined(CONFIG_ULP_COPROC_TYPE_RISCV)
+    ESP_ERROR_THROW(ulp_riscv_load_binary(ulp_pulse_counter_bin_start, size));
+    ESP_ERROR_THROW(ulp_riscv_run());
+#elif defined(CONFIG_ULP_COPROC_TYPE_LP_CORE)
+    ESP_ERROR_THROW(ulp_lp_core_load_binary(ulp_pulse_counter_bin_start, size));
+    ulp_lp_core_cfg_t cfg = {
+        .wakeup_source = ULP_LP_CORE_WAKEUP_SOURCE_HP_CPU,
+    };
+    ESP_ERROR_THROW(ulp_lp_core_run(&cfg));
+#endif
+
+    LOGTD(PULSE, "ULP pulse counter started with %" PRIu32 " channel(s)", ulpNextChannel);
+#endif
+}
+
+#ifdef CONFIG_ULP_COPROC_ENABLED
+
+std::shared_ptr<PulseCounter> PulseCounterManager::createUlp(const PulseCounterConfig& config) {
+    if (ulpNextChannel >= ULP_MAX_CHANNELS) {
+        throw std::runtime_error("PulseCounterManager: ULP channel limit exceeded");
+    }
+
+    auto gpio = config.pin->getGpio();
+
+    // Route pin to RTC/LP domain with pull-down.
+    ESP_ERROR_THROW(rtc_gpio_init(gpio));
+    ESP_ERROR_THROW(rtc_gpio_set_direction(gpio, RTC_GPIO_MODE_INPUT_ONLY));
+    ESP_ERROR_THROW(rtc_gpio_pulldown_en(gpio));
+    ESP_ERROR_THROW(rtc_gpio_pullup_dis(gpio));
+
+    uint32_t index = ulpNextChannel++;
+
+    uint32_t debounceUs = static_cast<uint32_t>(config.debounceTime.count());
+    ulp_gpio_num[index]    = static_cast<uint32_t>(gpio);
+    ulp_debounce_us[index] = debounceUs;
+    ulp_pulse_count[index] = 0;
+
+    LOGTD(PULSE, "Registered ULP pulse counter on pin %s (channel %" PRIu32 ", debounce %" PRIu32 " us)",
+        config.pin->getName().c_str(), index, debounceUs);
+
+    auto ulpCounter = std::shared_ptr<UlpPulseCounter>(new UlpPulseCounter(config.pin, index));
+    ulpCounters.push_back(ulpCounter);
+    return ulpCounter;
+}
+
+#endif    // CONFIG_ULP_COPROC_ENABLED
+
+std::shared_ptr<PulseCounter> PulseCounterManager::createGpio(const PulseCounterConfig& config) {
+    if (!gpioInitialized) {
+        gpioInitialized = true;
+
+        // Make sure we handle any state changes when the device wakes up due to a GPIO interrupt
+        esp_pm_sleep_cbs_register_config_t sleepCallbackConfig = {
+            .enter_cb = [](int64_t /*timeToSleepInUs*/, void* arg) {
+                auto* self = static_cast<PulseCounterManager*>(arg);
+                for (auto* counter : self->gpioCounters) {
+                    counter->handleGoingToLightSleep();
+                }
+                return ESP_OK; },
+            .exit_cb = [](int64_t /*timeSleptInUs*/, void* arg) {
+                auto* self = static_cast<PulseCounterManager*>(arg);
+                for (auto* counter : self->gpioCounters) {
+                    counter->handleWakingUpFromLightSleep();
+                }
+                return ESP_OK; },
+            .enter_cb_user_arg = this,
+            .exit_cb_user_arg = this,
+            .enter_cb_prior = 0,
+            .exit_cb_prior = 0,
+        };
+        ESP_ERROR_THROW(esp_pm_light_sleep_register_cbs(&sleepCallbackConfig));
+    }
+
+    auto counter = std::make_shared<GpioPulseCounter>(config.pin, config.debounceTime);
+
+    // Attach the ISR handler to the GPIO pin
+    ESP_ERROR_THROW(gpio_isr_handler_add(config.pin->getGpio(), handleGpioPulseCounterInterrupt, counter.get()));
+
+    // Keep the counter alive in the manager
+    gpioCounters.push_back(counter.get());
+    ulpCounters.push_back(counter);
+    return counter;
+}
+
+}    // namespace farmhub::kernel
